refactor double to BigDecimal

+++ src/main/java/calculator/Operation.java
@@ -2,6 +2,7 @@ package calculator;
 
 import visitor.Visitor;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
@@ -77,11 +78,12 @@ public abstract class Operation implements Expression
 
 	/**
 	 * Abstract method representing the actual binary arithmetic operation to compute
-	 * @param l	 first argument of the binary operation
-	 * @param r	second argument of the binary operation
-	 * @return	result of computing the binary operation
+	 *
+	 * @param l first argument of the binary operation
+	 * @param r second argument of the binary operation
+	 * @return result of computing the binary operation
 	 */
-   public abstract Double op(Double l, Double r);
+   public abstract BigDecimal op(BigDecimal l, BigDecimal r);
     // the operation itself is specified in the subclasses
 
 	/** Add more parameters to the existing list of parameters
